UdemyCourse/requirements.txt
pipdeptree
tiktoken
openai
python-dotenv
pypdf

# langchain

langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-qdrant

fastapi[standard]
aiofiles
uvicorn
